refactor(edit-engine): layer transaction execution helpers
