+++ packages/edit-engine/src/transaction.ts
@@ -9,50 +9,75 @@ import {
   NoConnectSchema,
   SchematicDocumentSchema,
   deriveStableId,
-  inverseTransformPoint,
-  powerDomainForNet,
-  powerNetNormalizations,
-  transformPoint,
 } from "@icm/model";
 import type {
   Annotation,
-  Orientation,
   Point,
   Rotation,
   RouteBranch,
-  RouteEndpoint,
   SchematicDocument,
 } from "@icm/model";
 import {
-  defaultInstanceLabelPlacement,
   endpointKey,
-  endpointBelongsToNet,
-  inferInstanceLabelSide,
-  instanceLabelRowOffset,
   isMosBulkRoute,
-  netEndpoints,
-  placeUprightInstanceLabel,
   resolveEndpointOutwardDirection,
   resolveEndpointPoint,
   resolveMosBulkConnection,
-  resolveSchematicStyleProfile,
-  visibleSymbolInkBounds,
 } from "@icm/derived";
-import { displayableInstanceValue } from "@icm/derived";
 import type { SymbolResolver } from "@icm/symbols";
-import { z } from "zod";
-
 import {
   EditTransactionSchema,
   type EditTransaction,
-  type SchematicEdit,
 } from "./edit-schema.js";
 import {
   buildOrthogonalEscapeRoute,
-  isOrthogonal,
   normalizeRouteGeometry,
 } from "./route-geometry-edit.js";
 import { resolveRouteEditPath } from "./route-operations.js";
+import {
+  followNetLabelsOnChangedRoutes,
+  followRouteMarkersOnChangedRoutes,
+  remapRouteMarkersAfterSplit,
+} from "./transaction-route-annotation-follow.js";
+import {
+  captureNetLabelRouteAnchors,
+  captureRouteMarkerAnchors,
+  closestRouteMarkerAnchor,
+  pointAtArcFraction,
+  type NetLabelRouteAnchor,
+  type RouteMarkerAnchor,
+} from "./transaction-route-annotations.js";
+import {
+  applyInstanceRouteFollow,
+  splitRoute,
+} from "./transaction-route-follow.js";
+import {
+  followAttachedAnnotations,
+  refreshInstanceValueAnnotation,
+  translateObjectAnchoredAnnotation,
+} from "./transaction-instance-annotations.js";
+import {
+  addEndpointToNet,
+  endpointOwnerNetId,
+  lockedLayoutOwner,
+  netEndpointGroups,
+  normalizePowerNets,
+  pointOnSegment,
+  replaceLayoutReference,
+  routeFromEdit,
+  routeIsProtected,
+  sameResolvedRoutePoints,
+  validateConnectableEndpoint,
+  validateNetLabelBinding,
+  validateRoute,
+} from "./transaction-routing.js";
+import {
+  gridAlignmentDiagnostics,
+  isHistoryEdit,
+  schemaDiagnostics,
+  snapPointToDocumentGrid,
+  transactionDismissesFlightlineGuidance,
+} from "./transaction-preflight.js";
 import {
   rejectTransaction,
   type EditDiagnostic,
@@ -65,1363 +90,6 @@ import {
 
 export * from "./edit-schema.js";
 export * from "./transaction-result.js";
-
-function schemaDiagnostics(error: z.ZodError, code: string): EditDiagnostic[] {
-  return error.issues.map((issue) => ({
-    code,
-    severity: "error" as const,
-    message: issue.message,
-    path: issue.path.map((segment) =>
-      typeof segment === "symbol" ? (segment.description ?? "symbol") : segment,
-    ),
-  }));
-}
-
-function gridAlignmentDiagnostics(
-  value: unknown,
-  grid: number,
-  path: ReadonlyArray<string | number> = [],
-): EditDiagnostic[] {
-  if (Array.isArray(value)) {
-    return value.flatMap((item, index) =>
-      gridAlignmentDiagnostics(item, grid, [...path, index]),
-    );
-  }
-  if (!value || typeof value !== "object") return [];
-  const record = value as Record<string, unknown>;
-  const diagnostics: EditDiagnostic[] = [];
-  if (typeof record.x === "number" && typeof record.y === "number") {
-    for (const axis of ["x", "y"] as const) {
-      const coordinate = record[axis] as number;
-      if (coordinate % grid === 0) continue;
-      diagnostics.push({
-        code: "GRID_ALIGNMENT",
-        severity: "error",
-        message: `Document page coordinates must align to grid ${grid}`,
-        path: [...path, axis],
-      });
-    }
-  }
-  for (const [key, child] of Object.entries(record)) {
-    diagnostics.push(...gridAlignmentDiagnostics(child, grid, [...path, key]));
-  }
-  return diagnostics;
-}
-
-function snapPointToDocumentGrid(point: Point, grid: number): Point {
-  return {
-    x: Math.round(point.x / grid) * grid,
-    y: Math.round(point.y / grid) * grid,
-  };
-}
-
-/**
- * Source synchronization and routing guidance have different lifetimes. An
- * imported Document starts with dashed guidance active; placing its initially
- * unplaced symbols or committing ordinary Wire must not dismiss the remaining
- * imported Nets. Deliberate presentation/geometry interventions do dismiss it
- * so a manually revised drawing is never repopulated with inferred guidance.
- */
-function transactionDismissesFlightlineGuidance(
-  document: SchematicDocument,
-  edits: readonly SchematicEdit[],
-): boolean {
-  if (!document.sourceBinding || document.flightlineGuidance === "dismissed") {
-    return false;
-  }
-  const isWireCommit = edits.some((edit) => edit.kind === "connect_endpoints");
-  return edits.some((edit) => {
-    switch (edit.kind) {
-      case "upsert_schematic_annotation":
-        return edit.annotation.kind === "net-label";
-      case "remove_schematic_annotation":
-        return (
-          document.annotations.find(
-            (annotation) => annotation.id === edit.annotationId,
-          )?.kind === "net-label"
-        );
-      case "move_instance":
-      case "rotate_instance":
-      case "mirror_instance":
-      case "align_instances":
-      case "move_junction":
-      case "cut_connection":
-      case "make_flightline":
-        return true;
-      case "set_route_points":
-      case "route_orthogonal":
-      case "add_junction":
-      case "attach_endpoint_to_route":
-        return !isWireCommit;
-      default:
-        return false;
-    }
-  });
-}
-
-function isHistoryEdit(
-  edit: SchematicEdit,
-): edit is Extract<SchematicEdit, { kind: "undo" | "redo" }> {
-  return edit.kind === "undo" || edit.kind === "redo";
-}
-
-function pointOnSegment(point: Point, from: Point, to: Point): boolean {
-  if (from.x === to.x) {
-    return (
-      point.x === from.x &&
-      point.y > Math.min(from.y, to.y) &&
-      point.y < Math.max(from.y, to.y)
-    );
-  }
-  if (from.y === to.y) {
-    return (
-      point.y === from.y &&
-      point.x > Math.min(from.x, to.x) &&
-      point.x < Math.max(from.x, to.x)
-    );
-  }
-  return false;
-}
-
-function routeIsProtected(route: RouteBranch): boolean {
-  return route.segmentModes.includes("locked");
-}
-
-function endpointOwnerNetId(
-  document: SchematicDocument,
-  endpoint: RouteEndpoint,
-): string | null {
-  switch (endpoint.kind) {
-    case "terminal":
-      return (
-        document.nets.find((net) =>
-          net.terminals.some(
-            (terminal) =>
-              terminal.instanceId === endpoint.instanceId &&
-              terminal.pinName === endpoint.pinName,
-          ),
-        )?.id ?? null
-      );
-    case "junction":
-      return (
-        document.junctions.find(
-          (junction) => junction.id === endpoint.junctionId,
-        )?.netId ?? null
-      );
-  }
-}
-
-function netEndpointGroups(
-  document: SchematicDocument,
-  netId: string,
-): string[][] {
-  const net = document.nets.find((candidate) => candidate.id === netId);
-  if (!net) return [];
-  const keys = netEndpoints(document, net).map(endpointKey);
-  const parent = new Map(keys.map((key) => [key, key]));
-  const find = (key: string): string => {
-    const current = parent.get(key);
-    if (!current) throw new Error(`Unknown Net endpoint ${key}`);
-    if (current === key) return key;
-    const root = find(current);
-    parent.set(key, root);
-    return root;
-  };
-  const union = (left: string, right: string): void => {
-    const leftRoot = find(left);
-    const rightRoot = find(right);
-    if (leftRoot === rightRoot) return;
-    const [first, second] = [leftRoot, rightRoot].sort((a, b) =>
-      a.localeCompare(b, "en"),
-    );
-    parent.set(second!, first!);
-  };
-  for (const route of document.routes.filter(
-    (candidate) => candidate.netId === netId,
-  )) {
-    union(endpointKey(route.from), endpointKey(route.to));
-  }
-  const grouped = new Map<string, string[]>();
-  for (const key of keys) {
-    const root = find(key);
-    const group = grouped.get(root) ?? [];
-    group.push(key);
-    grouped.set(root, group);
-  }
-  return [...grouped.values()]
-    .map((group) =>
-      group.sort((left, right) => left.localeCompare(right, "en")),
-    )
-    .sort((left, right) => left[0]!.localeCompare(right[0]!, "en"));
-}
-
-function validateConnectableEndpoint(
-  document: SchematicDocument,
-  endpoint: RouteEndpoint,
-  resolver: SymbolResolver | undefined,
-): string | null {
-  switch (endpoint.kind) {
-    case "terminal": {
-      const instance = document.instances.find(
-        (candidate) => candidate.id === endpoint.instanceId,
-      );
-      if (!instance) return `Instance does not exist: ${endpoint.instanceId}`;
-      if (!resolver) return "Terminal edits require a Symbol Resolver";
-      const symbol = resolver.resolve(
-        instance.symbolId,
-        instance.symbolVariantId,
-      );
-      if (
-        !symbol?.definition.pins.some((pin) => pin.name === endpoint.pinName)
-      ) {
-        return `Symbol pin does not exist: ${endpoint.instanceId}.${endpoint.pinName}`;
-      }
-      return null;
-    }
-    case "junction":
-      return document.junctions.some(
-        (junction) => junction.id === endpoint.junctionId,
-      )
-        ? null
-        : `Junction does not exist: ${endpoint.junctionId}`;
-  }
-}
-
-function validateNetLabelBinding(
-  document: SchematicDocument,
-  annotation: Annotation,
-): string | null {
-  if (annotation.kind !== "net-label") return null;
-  if (!annotation.netId) {
-    return `Net Label requires a Net identity: ${annotation.id}`;
-  }
-  return document.nets.some((net) => net.id === annotation.netId)
-    ? null
-    : `Net Label identity is not a Net: ${annotation.netId}`;
-}
-
-function addEndpointToNet(
-  document: SchematicDocument,
-  netId: string,
-  endpoint: RouteEndpoint,
-): void {
-  const net = document.nets.find((candidate) => candidate.id === netId)!;
-  if (endpoint.kind === "terminal") {
-    if (
-      !net.terminals.some(
-        (terminal) =>
-          terminal.instanceId === endpoint.instanceId &&
-          terminal.pinName === endpoint.pinName,
-      )
-    ) {
-      net.terminals.push({
-        instanceId: endpoint.instanceId,
-        pinName: endpoint.pinName,
-      });
-    }
-  }
-}
-
-function normalizePowerNets(
-  document: SchematicDocument,
-  changedObjectIds: Set<string>,
-): boolean {
-  let changed = false;
-  for (const normalization of powerNetNormalizations(document)) {
-    const net = document.nets.find(
-      (candidate) => candidate.id === normalization.netId,
-    )!;
-    let netChanged = false;
-    if (net.scope !== "global") {
-      net.scope = "global";
-      changed = true;
-      netChanged = true;
-    }
-    if (normalization.name && net.name !== normalization.name) {
-      net.name = normalization.name;
-      changed = true;
-      netChanged = true;
-    }
-    if (netChanged) changedObjectIds.add(net.id);
-  }
-  return changed;
-}
-
-function replaceLayoutReference(
-  objectIds: string[],
-  sourceId: string,
-  targetId: string,
-): string[] {
-  return [...new Set(objectIds.map((id) => (id === sourceId ? targetId : id)))];
-}
-
-function lockedLayoutOwner(
-  document: SchematicDocument,
-  objectId: string,
-): string | null {
-  return (
-    [...document.layoutGroups, ...document.constraints].find(
-      (item) => item.locked && item.objectIds.includes(objectId),
-    )?.id ?? null
-  );
-}
-
-function routeFromEdit(
-  edit: Extract<SchematicEdit, { kind: "set_route_points" }>,
-): RouteBranch {
-  return {
-    id: edit.routeId,
-    netId: edit.netId,
-    from: structuredClone(edit.from),
-    to: structuredClone(edit.to),
-    waypoints: structuredClone(edit.waypoints),
-    segmentModes: [...edit.segmentModes],
-    ...(edit.presentation ? { presentation: edit.presentation } : {}),
-  };
-}
-
-function validateRoute(
-  document: SchematicDocument,
-  route: RouteBranch,
-  resolver: SymbolResolver,
-): string | null {
-  if (route.segmentModes.length !== route.waypoints.length + 1) {
-    return `Route ${route.id} requires one segment mode per geometric segment`;
-  }
-  const net = document.nets.find((candidate) => candidate.id === route.netId);
-  if (!net) return `Route net does not exist: ${route.netId}`;
-  if (route.presentation === "power-rail" && powerDomainForNet(net) !== "vdd") {
-    return `Power rail ${route.id} must belong to a VDD Net`;
-  }
-  if (!endpointBelongsToNet(document, net, route.from)) {
-    return `Route from endpoint is not a member of ${route.netId}`;
-  }
-  if (!endpointBelongsToNet(document, net, route.to)) {
-    return `Route to endpoint is not a member of ${route.netId}`;
-  }
-  const polyline = resolveRouteEditPath(document, resolver, route);
-  if (!polyline) return `Route ${route.id} has an unresolved endpoint`;
-  if (!isOrthogonal(polyline.points)) {
-    return `Route ${route.id} must contain only non-zero orthogonal segments`;
-  }
-  for (const [endpoint, point, adjacent, mode] of [
-    [
-      route.from,
-      polyline.points[0]!,
-      polyline.points[1]!,
-      route.segmentModes[0],
-    ],
-    [
-      route.to,
-      polyline.points.at(-1)!,
-      polyline.points.at(-2)!,
-      route.segmentModes.at(-1),
-    ],
-  ] as const) {
-    if (endpoint.kind !== "terminal" || mode !== "escape") continue;
-    const outward = resolveEndpointOutwardDirection(
-      document,
-      resolver,
-      endpoint,
-    );
-    if (!outward) return `Route ${route.id} has an unresolved pin direction`;
-    const departure = { x: adjacent.x - point.x, y: adjacent.y - point.y };
-    if (departure.x * outward.x + departure.y * outward.y <= 0) {
-      return `Route ${route.id} escape segment must leave ${endpoint.instanceId}.${endpoint.pinName} outward`;
-    }
-  }
-  return null;
-}
-
-function sameResolvedRoutePoints(
-  left: readonly Point[] | null,
-  right: readonly Point[] | null,
-): boolean {
-  if (left === null || right === null) return left === right;
-  return (
-    left.length === right.length &&
-    left.every(
-      (point, index) =>
-        point.x === right[index]!.x && point.y === right[index]!.y,
-    )
-  );
-}
-
-interface NetLabelRouteAnchor {
-  annotationId: string;
-  routeId: string;
-  segmentIndex: number;
-  segmentCount: number;
-  t: number;
-  normalOffset: number;
-  arcFraction: number;
-}
-
-interface RouteMarkerAnchor {
-  annotationId: string;
-  routeId: string;
-  segmentIndex: number;
-  segmentCount: number;
-  t: number;
-  position: Point;
-  direction: Point;
-  routeStart: Point;
-  routeEnd: Point;
-}
-
-function closestRouteMarkerAnchor(
-  points: readonly Point[],
-  position: Point,
-  preferredDirection: Point,
-): { segmentIndex: number; t: number; distanceSquared: number } | null {
-  const candidates = points.slice(0, -1).flatMap((from, segmentIndex) => {
-    const to = points[segmentIndex + 1]!;
-    const dx = to.x - from.x;
-    const dy = to.y - from.y;
-    const lengthSquared = dx * dx + dy * dy;
-    if (lengthSquared === 0) return [];
-    const t = Math.max(
-      0,
-      Math.min(
-        1,
-        ((position.x - from.x) * dx + (position.y - from.y) * dy) /
-          lengthSquared,
-      ),
-    );
-    const anchor = { x: from.x + dx * t, y: from.y + dy * t };
-    const direction = { x: Math.sign(dx), y: Math.sign(dy) };
-    return [
-      {
-        segmentIndex,
-        t,
-        distanceSquared:
-          (position.x - anchor.x) ** 2 + (position.y - anchor.y) ** 2,
-        directionPenalty:
-          direction.x === preferredDirection.x &&
-          direction.y === preferredDirection.y
-            ? 0
-            : direction.x === -preferredDirection.x &&
-                direction.y === -preferredDirection.y
-              ? 1
-              : 2,
-      },
-    ];
-  });
-  const closest = candidates.sort(
-    (left, right) =>
-      left.distanceSquared - right.distanceSquared ||
-      left.directionPenalty - right.directionPenalty ||
-      left.segmentIndex - right.segmentIndex,
-  )[0];
-  return closest
-    ? {
-        segmentIndex: closest.segmentIndex,
-        t: closest.t,
-        distanceSquared: closest.distanceSquared,
-      }
-    : null;
-}
-
-function routeMarkerAttachment(annotation: Annotation) {
-  if (annotation.kind !== "route-marker") return null;
-  if (annotation.anchor.kind === "route") {
-    return {
-      routeId: annotation.anchor.routeId,
-      segmentIndex: annotation.anchor.segmentIndex,
-      t: annotation.anchor.t,
-      direction: annotation.anchor.direction,
-      normalOffset: annotation.anchor.normalOffset,
-    };
-  }
-  return null;
-}
-
-function closestRouteAnchor(
-  points: readonly Point[],
-  position: Point,
-):
-  | (Omit<NetLabelRouteAnchor, "annotationId" | "routeId" | "segmentCount"> & {
-      distanceSquared: number;
-    })
-  | null {
-  const lengths = points.slice(0, -1).map((from, index) => {
-    const to = points[index + 1]!;
-    return Math.hypot(to.x - from.x, to.y - from.y);
-  });
-  const totalLength = lengths.reduce((sum, length) => sum + length, 0);
-  if (totalLength === 0) return null;
-  let traversed = 0;
-  const candidates = lengths.flatMap((length, segmentIndex) => {
-    const from = points[segmentIndex]!;
-    const to = points[segmentIndex + 1]!;
-    if (length === 0) return [];
-    const dx = to.x - from.x;
-    const dy = to.y - from.y;
-    const t = Math.max(
-      0,
-      Math.min(
-        1,
-        ((position.x - from.x) * dx + (position.y - from.y) * dy) /
-          (length * length),
-      ),
-    );
-    const anchor = { x: from.x + dx * t, y: from.y + dy * t };
-    const delta = {
-      x: position.x - anchor.x,
-      y: position.y - anchor.y,
-    };
-    const candidate = {
-      segmentIndex,
-      t,
-      normalOffset: delta.x * (-dy / length) + delta.y * (dx / length),
-      arcFraction: (traversed + t * length) / totalLength,
-      distanceSquared: delta.x * delta.x + delta.y * delta.y,
-    };
-    traversed += length;
-    return [candidate];
-  });
-  return (
-    candidates.sort(
-      (left, right) =>
-        left.distanceSquared - right.distanceSquared ||
-        left.segmentIndex - right.segmentIndex,
-    )[0] ?? null
-  );
-}
-
-function captureNetLabelRouteAnchors(
-  document: SchematicDocument,
-  resolver: SymbolResolver,
-): NetLabelRouteAnchor[] {
-  const polylines = document.routes.flatMap((route) => {
-    const polyline = resolveRouteEditPath(document, resolver, route);
-    return polyline ? [{ route, polyline }] : [];
-  });
-  return document.annotations.flatMap((annotation) => {
-    const annotationAnchor = annotation.anchor;
-    if (
-      (annotation.kind !== "net-label" && annotation.kind !== "power-label") ||
-      annotationAnchor.kind !== "route"
-    ) {
-      return [];
-    }
-    const closest = polylines
-      .filter(({ route }) => route.id === annotationAnchor.routeId)
-      .flatMap(({ route, polyline }) => {
-        const anchor = closestRouteAnchor(
-          polyline.points,
-          annotationAnchor.fallbackPosition,
-        );
-        return anchor
-          ? [
-              {
-                ...anchor,
-                annotationId: annotation.id,
-                routeId: route.id,
-                segmentCount: polyline.points.length - 1,
-              },
-            ]
-          : [];
-      })
-      .sort(
-        (left, right) =>
-          left.distanceSquared - right.distanceSquared ||
-          left.routeId.localeCompare(right.routeId, "en"),
-      )[0];
-    if (!closest) return [];
-    const { distanceSquared: _distanceSquared, ...anchor } = closest;
-    return [anchor];
-  });
-}
-
-function captureRouteMarkerAnchors(
-  document: SchematicDocument,
-  resolver: SymbolResolver,
-): RouteMarkerAnchor[] {
-  return document.annotations.flatMap((annotation) => {
-    const attachment = routeMarkerAttachment(annotation);
-    if (!attachment) return [];
-    const route = document.routes.find(
-      (candidate) => candidate.id === attachment.routeId,
-    );
-    if (!route) return [];
-    const polyline = resolveRouteEditPath(document, resolver, route);
-    if (!polyline) return [];
-    const from = polyline.points[attachment.segmentIndex];
-    const to = polyline.points[attachment.segmentIndex + 1];
-    const routeStart = polyline.points[0];
-    const routeEnd = polyline.points.at(-1);
-    if (!from || !to || !routeStart || !routeEnd) return [];
-    return [
-      {
-        annotationId: annotation.id,
-        routeId: route.id,
-        segmentIndex: attachment.segmentIndex,
-        segmentCount: polyline.points.length - 1,
-        t: attachment.t,
-        position: {
-          x: from.x + (to.x - from.x) * attachment.t,
-          y: from.y + (to.y - from.y) * attachment.t,
-        },
-        direction: { x: Math.sign(to.x - from.x), y: Math.sign(to.y - from.y) },
-        routeStart,
-        routeEnd,
-      },
-    ];
-  });
-}
-
-function pointAtArcFraction(
-  points: readonly Point[],
-  fraction: number,
-): { segmentIndex: number; t: number } | null {
-  const lengths = points.slice(0, -1).map((from, index) => {
-    const to = points[index + 1]!;
-    return Math.hypot(to.x - from.x, to.y - from.y);
-  });
-  const total = lengths.reduce((sum, length) => sum + length, 0);
-  if (total === 0) return null;
-  const target = Math.max(0, Math.min(1, fraction)) * total;
-  let traversed = 0;
-  for (const [segmentIndex, length] of lengths.entries()) {
-    if (length === 0) continue;
-    if (traversed + length >= target || segmentIndex === lengths.length - 1) {
-      return {
-        segmentIndex,
-        t: Math.max(0, Math.min(1, (target - traversed) / length)),
-      };
-    }
-    traversed += length;
-  }
-  return null;
-}
-
-function followNetLabelsOnChangedRoutes(
-  draft: SchematicDocument,
-  resolver: SymbolResolver,
-  anchors: readonly NetLabelRouteAnchor[],
-  changedRouteIds: ReadonlySet<string>,
-  changedObjectIds: Set<string>,
-): void {
-  for (const captured of anchors) {
-    if (
-      !changedRouteIds.has(captured.routeId) ||
-      changedObjectIds.has(captured.annotationId)
-    ) {
-      continue;
-    }
-    const annotation = draft.annotations.find(
-      (candidate) => candidate.id === captured.annotationId,
-    );
-    const route = draft.routes.find(
-      (candidate) => candidate.id === captured.routeId,
-    );
-    if (!annotation || annotation.kind !== "net-label" || !route) continue;
-    const polyline = resolveRouteEditPath(draft, resolver, route);
-    if (!polyline) continue;
-    const segmentCount = polyline.points.length - 1;
-    const attachment =
-      segmentCount === captured.segmentCount &&
-      captured.segmentIndex < segmentCount
-        ? { segmentIndex: captured.segmentIndex, t: captured.t }
-        : pointAtArcFraction(polyline.points, captured.arcFraction);
-    if (!attachment) continue;
-    const from = polyline.points[attachment.segmentIndex]!;
-    const to = polyline.points[attachment.segmentIndex + 1]!;
-    const dx = to.x - from.x;
-    const dy = to.y - from.y;
-    const length = Math.hypot(dx, dy);
-    if (length === 0) continue;
-    const anchor = {
-      x: from.x + dx * attachment.t,
-      y: from.y + dy * attachment.t,
-    };
-    const normal = { x: -dy / length, y: dx / length };
-    const offset = {
-      x: normal.x * captured.normalOffset,
-      y: normal.y * captured.normalOffset,
-    };
-    if (annotation.anchor.kind !== "route") continue;
-    annotation.anchor = {
-      ...annotation.anchor,
-      segmentIndex: attachment.segmentIndex,
-      t: attachment.t,
-      normalOffset: Math.round(offset.x * normal.x + offset.y * normal.y),
-      fallbackPosition: snapPointToDocumentGrid(
-        { x: anchor.x + offset.x, y: anchor.y + offset.y },
-        draft.presentation.grid,
-      ),
-    };
-    changedObjectIds.add(annotation.id);
-  }
-}
-
-function followRouteMarkersOnChangedRoutes(
-  draft: SchematicDocument,
-  resolver: SymbolResolver,
-  anchors: readonly RouteMarkerAnchor[],
-  changedRouteIds: ReadonlySet<string>,
-  changedObjectIds: Set<string>,
-): void {
-  for (const captured of anchors) {
-    if (
-      !changedRouteIds.has(captured.routeId) ||
-      changedObjectIds.has(captured.annotationId)
-    ) {
-      continue;
-    }
-    const annotation = draft.annotations.find(
-      (candidate) => candidate.id === captured.annotationId,
-    );
-    const route = draft.routes.find(
-      (candidate) => candidate.id === captured.routeId,
-    );
-    if (!annotation || annotation.kind !== "route-marker" || !route) continue;
-    const polyline = resolveRouteEditPath(draft, resolver, route);
-    if (!polyline) continue;
-    const segmentCount = polyline.points.length - 1;
-    let attachment =
-      segmentCount === captured.segmentCount &&
-      captured.segmentIndex < segmentCount
-        ? { segmentIndex: captured.segmentIndex, t: captured.t }
-        : null;
-    if (!attachment) {
-      const nextStart = polyline.points[0]!;
-      const nextEnd = polyline.points.at(-1)!;
-      const startDelta = {
-        x: nextStart.x - captured.routeStart.x,
-        y: nextStart.y - captured.routeStart.y,
-      };
-      const endDelta = {
-        x: nextEnd.x - captured.routeEnd.x,
-        y: nextEnd.y - captured.routeEnd.y,
-      };
-      const expectedPosition =
-        startDelta.x === endDelta.x && startDelta.y === endDelta.y
-          ? {
-              x: captured.position.x + startDelta.x,
-              y: captured.position.y + startDelta.y,
-            }
-          : captured.position;
-      const closest = closestRouteMarkerAnchor(
-        polyline.points,
-        expectedPosition,
-        captured.direction,
-      );
-      attachment = closest
-        ? { segmentIndex: closest.segmentIndex, t: closest.t }
-        : null;
-    }
-    if (!attachment) continue;
-    const from = polyline.points[attachment.segmentIndex]!;
-    const to = polyline.points[attachment.segmentIndex + 1]!;
-    const position = snapPointToDocumentGrid(
-      {
-        x: from.x + (to.x - from.x) * attachment.t,
-        y: from.y + (to.y - from.y) * attachment.t,
-      },
-      draft.presentation.grid,
-    );
-    if (annotation.anchor.kind === "route") {
-      annotation.anchor = {
-        ...annotation.anchor,
-        segmentIndex: attachment.segmentIndex,
-        t: attachment.t,
-        fallbackPosition: position,
-      };
-    }
-    changedObjectIds.add(annotation.id);
-  }
-}
-
-function remapRouteMarkersAfterSplit(
-  draft: SchematicDocument,
-  resolver: SymbolResolver,
-  anchors: readonly RouteMarkerAnchor[],
-  splitRouteIds: readonly string[],
-  changedObjectIds: Set<string>,
-): void {
-  for (const captured of anchors) {
-    const closest = splitRouteIds
-      .flatMap((routeId) => {
-        const route = draft.routes.find(
-          (candidate) => candidate.id === routeId,
-        );
-        const polyline = route
-          ? resolveRouteEditPath(draft, resolver, route)
-          : null;
-        if (!route || !polyline) return [];
-        const attachment = closestRouteMarkerAnchor(
-          polyline.points,
-          captured.position,
-          captured.direction,
-        );
-        return attachment ? [{ route, polyline, attachment }] : [];
-      })
-      .sort(
-        (left, right) =>
-          left.attachment.distanceSquared - right.attachment.distanceSquared ||
-          left.route.id.localeCompare(right.route.id, "en"),
-      )[0];
-    if (!closest) continue;
-    const annotation = draft.annotations.find(
-      (candidate) => candidate.id === captured.annotationId,
-    );
-    if (!annotation || annotation.kind !== "route-marker") continue;
-    const { segmentIndex, t } = closest.attachment;
-    const from = closest.polyline.points[segmentIndex]!;
-    const to = closest.polyline.points[segmentIndex + 1]!;
-    const position = snapPointToDocumentGrid(
-      {
-        x: from.x + (to.x - from.x) * t,
-        y: from.y + (to.y - from.y) * t,
-      },
-      draft.presentation.grid,
-    );
-    if (annotation.anchor.kind === "route") {
-      annotation.anchor = {
-        ...annotation.anchor,
-        routeId: closest.route.id,
-        segmentIndex,
-        t,
-        fallbackPosition: position,
-      };
-    }
-    changedObjectIds.add(annotation.id);
-  }
-}
-
-function splitRoute(
-  document: SchematicDocument,
-  route: RouteBranch,
-  splitEndpoint: RouteEndpoint,
-  position: Point,
-  firstRouteId: string,
-  secondRouteId: string,
-  segmentIndex: number,
-  resolver: SymbolResolver,
-): { first: RouteBranch; second: RouteBranch } | string {
-  const polyline = resolveRouteEditPath(document, resolver, route);
-  if (!polyline) return `Route ${route.id} has an unresolved endpoint`;
-  if (segmentIndex >= polyline.points.length - 1) {
-    return `Route split segment is out of range: ${segmentIndex}`;
-  }
-  const vertexIndex = polyline.points.findIndex(
-    (point, index) =>
-      index > 0 &&
-      index < polyline.points.length - 1 &&
-      point.x === position.x &&
-      point.y === position.y,
-  );
-  if (vertexIndex > 0) {
-    // A manual orthogonal bend is already a geometric vertex, not a point in
-    // the interior of either adjoining segment. Splitting it through the
-    // ordinary path would introduce a zero-length segment and is rejected by
-    // route validation. Partition the existing polyline at the vertex instead.
-    const firstNormalized = normalizeRouteGeometry(
-      polyline.points.slice(0, vertexIndex + 1),
-      route.segmentModes.slice(0, vertexIndex),
-    );
-    const secondNormalized = normalizeRouteGeometry(
-      polyline.points.slice(vertexIndex),
-      route.segmentModes.slice(vertexIndex),
-    );
-    return {
-      first: {
-        id: firstRouteId,
-        netId: route.netId,
-        from: structuredClone(route.from),
-        to: structuredClone(splitEndpoint),
-        waypoints: firstNormalized.points.slice(1, -1),
-        segmentModes: firstNormalized.segmentModes,
-        ...(route.presentation ? { presentation: route.presentation } : {}),
-      },
-      second: {
-        id: secondRouteId,
-        netId: route.netId,
-        from: structuredClone(splitEndpoint),
-        to: structuredClone(route.to),
-        waypoints: secondNormalized.points.slice(1, -1),
-        segmentModes: secondNormalized.segmentModes,
-        ...(route.presentation ? { presentation: route.presentation } : {}),
-      },
-    };
-  }
-  const segmentFrom = polyline.points[segmentIndex]!;
-  const segmentTo = polyline.points[segmentIndex + 1]!;
-  if (!pointOnSegment(position, segmentFrom, segmentTo)) {
-    return `Junction position is not inside route segment ${segmentIndex}`;
-  }
-  const firstNormalized = normalizeRouteGeometry(
-    [...polyline.points.slice(0, segmentIndex + 1), position],
-    route.segmentModes.slice(0, segmentIndex + 1),
-  );
-  const secondNormalized = normalizeRouteGeometry(
-    [position, ...polyline.points.slice(segmentIndex + 1)],
-    [
-      route.segmentModes[segmentIndex]!,
-      ...route.segmentModes.slice(segmentIndex + 1),
-    ],
-  );
-  return {
-    first: {
-      id: firstRouteId,
-      netId: route.netId,
-      from: structuredClone(route.from),
-      to: structuredClone(splitEndpoint),
-      waypoints: firstNormalized.points.slice(1, -1),
-      segmentModes: firstNormalized.segmentModes,
-      ...(route.presentation ? { presentation: route.presentation } : {}),
-    },
-    second: {
-      id: secondRouteId,
-      netId: route.netId,
-      from: structuredClone(splitEndpoint),
-      to: structuredClone(route.to),
-      waypoints: secondNormalized.points.slice(1, -1),
-      segmentModes: secondNormalized.segmentModes,
-      ...(route.presentation ? { presentation: route.presentation } : {}),
-    },
-  };
-}
-
-function samePoint(left: Point, right: Point): boolean {
-  return left.x === right.x && left.y === right.y;
-}
-
-function endpointBelongsToInstance(
-  endpoint: RouteEndpoint,
-  instanceId: string,
-): boolean {
-  return endpoint.kind === "terminal" && endpoint.instanceId === instanceId;
-}
-
-/**
- * Move one resolved terminal endpoint while preserving the axis of its
- * adjacent persisted segment. This changes geometry only; it never changes
- * Route topology or connectivity.
- */
-function followRouteEndpoint(
-  routeId: string,
-  points: Point[],
-  modes: RouteBranch["segmentModes"],
-  side: "from" | "to",
-  oldPoint: Point,
-  newPoint: Point,
-  outward: Point | null,
-): void {
-  if (samePoint(oldPoint, newPoint)) return;
-  const mode = side === "from" ? modes[0] : modes.at(-1);
-  if (mode === "locked" || mode === "trunk") {
-    throw new Error(`Route ${routeId} has a protected adjacent segment`);
-  }
-  const endpointIndex = side === "from" ? 0 : points.length - 1;
-  points[endpointIndex] = { ...newPoint };
-  if (points.length === 2) return;
-
-  const neighborIndex = side === "from" ? 1 : points.length - 2;
-  const neighbor = points[neighborIndex]!;
-  const oldNeighbor = { ...neighbor };
-  if (mode === "escape" && outward) {
-    const escapeLength =
-      Math.abs(oldNeighbor.x - oldPoint.x) +
-      Math.abs(oldNeighbor.y - oldPoint.y);
-    neighbor.x = newPoint.x + outward.x * escapeLength;
-    neighbor.y = newPoint.y + outward.y * escapeLength;
-
-    const nextIndex = side === "from" ? neighborIndex + 1 : neighborIndex - 1;
-    const next = points[nextIndex]!;
-    if (neighbor.x !== next.x && neighbor.y !== next.y) {
-      // Turn away from the rotated/mirrored escape before reconnecting to the
-      // unchanged body. Choosing the perpendicular axis avoids a collinear
-      // U-turn that normalization would collapse back toward the pin.
-      const bridge =
-        outward.x !== 0
-          ? { x: neighbor.x, y: next.y }
-          : { x: next.x, y: neighbor.y };
-      const bridgeModeIndex = side === "from" ? 1 : modes.length - 2;
-      const bridgeMode = modes[bridgeModeIndex] ?? "auto";
-      points.splice(side === "from" ? nextIndex : neighborIndex, 0, bridge);
-      modes.splice(bridgeModeIndex, 1, bridgeMode, bridgeMode);
-    }
-    return;
-  }
-  if (oldPoint.x === neighbor.x && oldPoint.y !== neighbor.y) {
-    neighbor.x = newPoint.x;
-  } else if (oldPoint.y === neighbor.y && oldPoint.x !== neighbor.x) {
-    neighbor.y = newPoint.y;
-  } else {
-    throw new Error(`Route ${routeId} has invalid endpoint geometry`);
-  }
-}
-
-/**
- * Apply topology-preserving Route geometry after any instance placement
- * transform. The caller supplies the pre-edit snapshot and the transformed
- * draft, making move/rotate/mirror share one behavior at the transaction
- * boundary.
- */
-function applyInstanceRouteFollow(
-  draft: SchematicDocument,
-  originalDocument: SchematicDocument,
-  resolver: SymbolResolver,
-  instanceId: string,
-  explicitlyAuthoredRouteIds: ReadonlySet<string>,
-): string[] {
-  const changed: string[] = [];
-  for (const originalRoute of originalDocument.routes) {
-    if (explicitlyAuthoredRouteIds.has(originalRoute.id)) continue;
-    const movesFrom = endpointBelongsToInstance(originalRoute.from, instanceId);
-    const movesTo = endpointBelongsToInstance(originalRoute.to, instanceId);
-    if (!movesFrom && !movesTo) continue;
-
-    const route = draft.routes.find(
-      (candidate) => candidate.id === originalRoute.id,
-    );
-    const original = resolveRouteEditPath(
-      originalDocument,
-      resolver,
-      originalRoute,
-    );
-    const newFrom = route
-      ? resolveEndpointPoint(draft, resolver, route.from)
-      : null;
-    const newTo = route
-      ? resolveEndpointPoint(draft, resolver, route.to)
-      : null;
-    if (!route || !original || !newFrom || !newTo) continue;
-
-    const points = original.points.map((point) => ({ ...point }));
-    const modes = [...original.segmentModes];
-    try {
-      if (movesFrom) {
-        followRouteEndpoint(
-          route.id,
-          points,
-          modes,
-          "from",
-          original.points[0]!,
-          newFrom,
-          resolveEndpointOutwardDirection(draft, resolver, route.from),
-        );
-      }
-      if (movesTo) {
-        followRouteEndpoint(
-          route.id,
-          points,
-          modes,
-          "to",
-          original.points.at(-1)!,
-          newTo,
-          resolveEndpointOutwardDirection(draft, resolver, route.to),
-        );
-      }
-    } catch {
-      // Protected or otherwise non-followable geometry remains unchanged;
-      // final validation rejects the transaction and names the affected Route.
-      // Routes explicitly authored anywhere in this transaction were skipped
-      // above, so their edit is the sole geometry authority.
-      continue;
-    }
-
-    if (points.length === 2 && newFrom.x !== newTo.x && newFrom.y !== newTo.y) {
-      const originallyVertical =
-        original.points[0]!.x === original.points[1]!.x;
-      points.splice(
-        1,
-        0,
-        originallyVertical
-          ? { x: newFrom.x, y: newTo.y }
-          : { x: newTo.x, y: newFrom.y },
-      );
-      const mode = modes[0] ?? "manual";
-      modes.splice(0, 1, mode, mode);
-    }
-
-    const normalized = normalizeRouteGeometry(points, modes);
-    if (!isOrthogonal(normalized.points)) continue;
-    route.waypoints = normalized.points.slice(1, -1);
-    route.segmentModes = normalized.segmentModes;
-    changed.push(route.id);
-  }
-  return changed.sort((left, right) => left.localeCompare(right, "en"));
-}
-
-function translateObjectAnchoredAnnotation(
-  annotation: Annotation,
-  objectId: string,
-  delta: Point,
-): void {
-  if (
-    annotation.anchor.kind === "object" &&
-    annotation.anchor.objectId === objectId
-  ) {
-    annotation.anchor.fallbackPosition = {
-      x: annotation.anchor.fallbackPosition.x + delta.x,
-      y: annotation.anchor.fallbackPosition.y + delta.y,
-    };
-  }
-}
-
-/** The upright text slot an annotation occupies beside its instance. */
-function instanceAnnotationSlot(
-  annotation: Annotation,
-): "reference" | "value" | null {
-  if (annotation.anchor.kind !== "object") return null;
-  if (annotation.kind === "instance-label") return "reference";
-  if (annotation.kind === "instance-value") return "value";
-  return null;
-}
-
-/**
- * Re-project the machine-managed Value annotation after a parameter edit.
- * A Value whose text no longer equals the previous projection is treated as
- * hand-edited and left untouched (the clipboard instance-label precedent).
- * When the new projection is undisplayable the annotation keeps its text but
- * hides itself instead of showing stale electrical claims; the editor's show
- * toggle re-projects fresh content.
- */
-function refreshInstanceValueAnnotation(
-  draft: SchematicDocument,
-  before: SchematicDocument["instances"][number],
-  instanceId: string,
-  changedObjectIds: Set<string>,
-): void {
-  const instance = draft.instances.find(
-    (candidate) => candidate.id === instanceId,
-  );
-  if (!instance) return;
-  const previous = displayableInstanceValue(before);
-  if (previous.kind !== "displayable") return;
-  for (const annotation of draft.annotations) {
-    if (
-      annotation.kind !== "instance-value" ||
-      annotation.anchor.kind !== "object" ||
-      annotation.anchor.objectId !== instanceId
-    ) {
-      continue;
-    }
-    if (
-      JSON.stringify(annotation.content) !== JSON.stringify(previous.content)
-    ) {
-      continue;
-    }
-    const next = displayableInstanceValue(instance);
-    if (next.kind === "displayable") {
-      annotation.content = structuredClone(next.content);
-    } else {
-      annotation.visible = false;
-    }
-    changedObjectIds.add(annotation.id);
-  }
-}
-
-/**
- * A reference or value label is renderer-managed only while it exactly agrees
- * with the current canonical default for its slot. A user-moved label remains
- * an authored object-relative vector and must not be pulled back onto the
- * automatic side when its instance is rotated or mirrored.
- */
-function isCanonicalInstanceLabel(
-  annotation: Annotation,
-  instance: SchematicDocument["instances"][number],
-  resolved: NonNullable<ReturnType<SymbolResolver["resolve"]>>,
-  document: SchematicDocument,
-  oldPosition: Point,
-  oldOrientation: Orientation,
-): boolean {
-  const slot = instanceAnnotationSlot(annotation);
-  if (!slot || annotation.anchor.kind !== "object") {
-    return false;
-  }
-  const placement = { position: oldPosition, ...oldOrientation };
-  const expected = defaultInstanceLabelPlacement(
-    { ...instance, placement },
-    resolved,
-    resolveSchematicStyleProfile(document.presentation.styleProfileId),
-    document.presentation.grid,
-    slot,
-  );
-  if (!expected) return false;
-  const visiblePosition = {
-    x: oldPosition.x + annotation.anchor.localOffset.x,
-    y: oldPosition.y + annotation.anchor.localOffset.y,
-  };
-  return (
-    annotation.alignment === expected.alignment &&
-    visiblePosition.x === expected.position.x &&
-    visiblePosition.y === expected.position.y &&
-    annotation.anchor.fallbackPosition.x === expected.position.x &&
-    annotation.anchor.fallbackPosition.y === expected.position.y
-  );
-}
-
-function followAttachedAnnotations(
-  draft: SchematicDocument,
-  instanceId: string,
-  oldPosition: Point,
-  oldOrientation: Orientation,
-  newPosition: Point,
-  newOrientation: Orientation,
-  changedObjectIds: Set<string>,
-  resolver?: SymbolResolver,
-): void {
-  const isPureTranslation =
-    oldOrientation.rotation === newOrientation.rotation &&
-    oldOrientation.mirror === newOrientation.mirror;
-  if (isPureTranslation) {
-    const delta = {
-      x: newPosition.x - oldPosition.x,
-      y: newPosition.y - oldPosition.y,
-    };
-    for (const annotation of draft.annotations) {
-      if (
-        annotation.anchor.kind !== "object" ||
-        annotation.anchor.objectId !== instanceId
-      ) {
-        continue;
-      }
-      translateObjectAnchoredAnnotation(annotation, instanceId, delta);
-      changedObjectIds.add(annotation.id);
-    }
-    return;
-  }
-
-  const directionForRotation = (rotation: Rotation): Point => {
-    switch (rotation) {
-      case 0:
-        return { x: 1, y: 0 };
-      case 90:
-        return { x: 0, y: 1 };
-      case 180:
-        return { x: -1, y: 0 };
-      case 270:
-        return { x: 0, y: -1 };
-    }
-  };
-  const rotationForDirection = (direction: Point): Rotation => {
-    if (direction.x > 0) return 0;
-    if (direction.y > 0) return 90;
-    if (direction.x < 0) return 180;
-    return 270;
-  };
-  const origin = { x: 0, y: 0 };
-  const instance = draft.instances.find(
-    (candidate) => candidate.id === instanceId,
-  );
-  const resolved = instance
-    ? resolver?.resolve(instance.symbolId, instance.symbolVariantId)
-    : undefined;
-  for (const annotation of draft.annotations) {
-    if (
-      annotation.anchor.kind !== "object" ||
-      annotation.anchor.objectId !== instanceId
-    ) {
-      continue;
-    }
-    const visiblePosition = {
-      x: oldPosition.x + annotation.anchor.localOffset.x,
-      y: oldPosition.y + annotation.anchor.localOffset.y,
-    };
-    const local = inverseTransformPoint(
-      visiblePosition,
-      oldPosition,
-      oldOrientation,
-    );
-    const transformedAnchor = transformPoint(
-      local,
-      newPosition,
-      newOrientation,
-    );
-    let position = transformedAnchor;
-    let transformedAlignment: "start" | "middle" | "end" | null = null;
-    const slot = instanceAnnotationSlot(annotation);
-    if (
-      slot !== null &&
-      instance &&
-      resolved &&
-      isCanonicalInstanceLabel(
-        annotation,
-        instance,
-        resolved,
-        draft,
-        oldPosition,
-        oldOrientation,
-      )
-    ) {
-      const styleProfile = resolveSchematicStyleProfile(
-        draft.presentation.styleProfileId,
-      );
-      // Upright rows stack along world y regardless of orientation, so the
-      // value slot's row offset is stripped from the recovered anchor in world
-      // space before side inference and re-added by the upright placer.
-      const rowOffset =
-        slot === "value"
-          ? instanceLabelRowOffset(styleProfile, draft.presentation.grid)
-          : 0;
-      const slotAnchor = rowOffset
-        ? { x: visiblePosition.x, y: visiblePosition.y - rowOffset }
-        : visiblePosition;
-      const slotLocal = inverseTransformPoint(
-        slotAnchor,
-        oldPosition,
-        oldOrientation,
-      );
-      const localSide = inferInstanceLabelSide(
-        slotLocal,
-        visibleSymbolInkBounds(resolved),
-      );
-      if (localSide) {
-        try {
-          const placement = placeUprightInstanceLabel(
-            instance,
-            resolved,
-            styleProfile,
-            slotLocal,
-            localSide,
-            draft.presentation.grid,
-            annotation.sizeScale,
-            rowOffset,
-          );
-          if (placement) {
-            position = placement.position;
-            transformedAlignment = placement.alignment;
-          }
-        } catch {
-          // Keep the rigid semantic transform for a legacy/unknown profile;
-          // formal rendering reports the invalid profile separately.
-        }
-      }
-    }
-    annotation.anchor = {
-      ...annotation.anchor,
-      // Object anchors resolve localOffset directly in world space. Persist
-      // the reflowed upright glyph baseline without a second grid snap. The
-      // label placer already performed the one authoritative grid snap;
-      // re-snapping a recovered anchor is what previously accumulated drift.
-      localOffset: {
-        x: position.x - newPosition.x,
-        y: position.y - newPosition.y,
-      },
-      fallbackPosition: position,
-    };
-    if (slot !== null) {
-      annotation.rotation = 0;
-      annotation.alignment = transformedAlignment ?? annotation.alignment;
-    } else {
-      const oldDirection = directionForRotation(annotation.rotation);
-      const localDirection = inverseTransformPoint(
-        oldDirection,
-        origin,
-        oldOrientation,
-      );
-      annotation.rotation = rotationForDirection(
-        transformPoint(localDirection, origin, newOrientation),
-      );
-    }
-    changedObjectIds.add(annotation.id);
-  }
-}
 
 /**
  * Ensure the ADR 0010 drafting layer exists on a draft Document. It is
